LibArchive: Move method implementations away from header
